+++ mlgames/QTrainer.cpp
@@ -14,9 +14,9 @@ QTrainer::QTrainer(
 
 
 void QTrainer::train_step(
-        size_t count_samples, size_t num_state, const std::vector<int> &old_states_,
+        size_t count_samples, size_t num_state, const std::vector<double> &old_states_,
         size_t num_action, const std::vector<int> &actions_, const std::vector<int> &rewards_,
-        const std::vector<int> &new_states_, const std::vector<bool> &dones_)
+        const std::vector<double> &new_states_, const std::vector<bool> &dones_)
 {
     const torch::Tensor old_states = torch::tensor(old_states_, torch::kInt)
                                              .reshape({(long) count_samples, (long) num_state})
