+++ mlgames/include/mlgames/Agent.h
@@ -69,10 +69,10 @@ public:
 
     size_t num_state;
     size_t num_action;
-    std::vector<int> state;
+    std::vector<double> state;
     std::vector<int> action;
     int reward;
-    std::vector<int> next_state;
+    std::vector<double> next_state;
     bool game_over;
 };
 
@@ -84,18 +84,18 @@ public:
     Agent(Linear_QNet *model_, QTrainer *trainer_, c10::DeviceType device_);
 
     std::vector<int> get_action(
-            const std::vector<int> &state, size_t num_state, size_t num_action,
+            const std::vector<double> &state, size_t num_state, size_t num_action,
             size_t count_samples) const;
 
     void train_short_memory(
-            size_t num_state, size_t num_action, const std::vector<int> &state,
-            const std::vector<int> &action, int reward, const std::vector<int> &next_state,
+            size_t num_state, size_t num_action, const std::vector<double> &state,
+            const std::vector<int> &action, int reward, const std::vector<double> &next_state,
             bool game_over) const;
     void train_long_memory();
 
     void remember(
-            size_t num_state, size_t num_action, const std::vector<int> &state,
-            const std::vector<int> &action, int reward, const std::vector<int> &next_state,
+            size_t num_state, size_t num_action, const std::vector<double> &state,
+            const std::vector<int> &action, int reward, const std::vector<double> &next_state,
             bool game_over);
 
     int number_of_games = 0;
@@ -108,5 +108,5 @@ private:
     LimitedDeque<MemoryData> memory_deque{MAX_MEMORY};
 
     std::vector<int>
-    get_play(const std::vector<int> &state, size_t num_state, size_t count_samples) const;
+    get_play(const std::vector<double> &state, size_t num_state, size_t count_samples) const;
 };
